Move Exceptions to file_operations module
