+++ app/src/main/java/com/amaze/filemanager/database/CloudHandler.java
@@ -23,7 +23,7 @@ package com.amaze.filemanager.database;
 import java.util.List;
 
 import com.amaze.filemanager.database.models.explorer.CloudEntry;
-import com.amaze.filemanager.exceptions.CloudPluginException;
+import com.amaze.filemanager.file_operations.exceptions.CloudPluginException;
 import com.amaze.filemanager.file_operations.filesystem.OpenMode;
 import com.amaze.filemanager.ui.fragments.CloudSheetFragment;
