+++ app/src/main/java/com/amaze/filemanager/filesystem/root/CopyFilesCommand.kt
@@ -20,7 +20,7 @@
 
 package com.amaze.filemanager.filesystem.root
 
-import com.amaze.filemanager.exceptions.ShellNotRunningException
+import com.amaze.filemanager.file_operations.exceptions.ShellNotRunningException
 import com.amaze.filemanager.filesystem.RootHelper
 import com.amaze.filemanager.filesystem.root.MountPathCommand.mountPath
 import com.amaze.filemanager.filesystem.root.base.IRootCommand
