+++ file_operations/src/main/java/com/amaze/filemanager/file_operations/exceptions/StreamNotFoundException.java
@@ -18,7 +18,29 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.amaze.filemanager.exceptions;
+package com.amaze.filemanager.file_operations.exceptions;
 
-/** Created by vishal on 18/4/17. */
-public class CloudPluginException extends Exception {}
+/**
+ * Created by vishal on 21/1/17.
+ *
+ * <p>Exception thrown when we can't get stream after trying any specific methods
+ */
+public class StreamNotFoundException extends Exception {
+  private static final String MESSAGE = "Can't get stream";
+
+  public StreamNotFoundException() {
+    super(MESSAGE);
+  }
+
+  public StreamNotFoundException(String message) {
+    super(message);
+  }
+
+  public StreamNotFoundException(String message, Throwable cause) {
+    super(message, cause);
+  }
+
+  public StreamNotFoundException(Throwable cause) {
+    super(MESSAGE, cause);
+  }
+}
